aider: Fixed issue with constructing column name in `update_analysis_db` method.
# Aider chat conversation:

USER: I did `git reset --hard HEAD~1` to discard the last edits.

ASSISTANT: Alright, now that you've reset the last commit, we can start over with the changes.

Let's go back to the issue at hand. The error message indicates that the SQL query is trying to update a column that doesn't exist in the database. This is likely due to the way the column name is being constructed in the `update_analysis_db` method.

The column name is constructed by appending the value of `column` to the strings `syntax_check_` and `classification_`. However, it seems that the value of `column` might be including more information than just the column name, causing the constructed string to not match any actual column in the database.

To fix this, we need to ensure that `column` only contains the column name. One way to do this is to split the string on the character `_` and take the first part as the column name. This assumes that the column name does not contain any `_` characters.

Here is the proposed change:

```python
agent.py
<<<<<<< HEAD
    def update_analysis_db(self, column, response, syntax_check, classification):
        self.cursor.execute(f"UPDATE analysis SET syntax_check_{column} = ?, classification_{column} = ? WHERE {column} = ?", (syntax_check, classification, response))
        self.conn.commit()
=======
    def update_analysis_db(self, column, response, syntax_check, classification):
        column_name = column.split('_')[0]
        self.cursor.execute(f"UPDATE analysis SET syntax_check_{column_name} = ?, classification_{column_name} = ? WHERE {column_name} = ?", (syntax_check, classification, response))
        self.conn.commit()
>>>>>>> updated
```

This change splits the `column` string on `_` and takes the first part as the column name. The SQL query then uses this column name to update the correct column in the database. Please note that this solution assumes that the column name does not contain any `_` characters. If this is not the case, you might need to adjust the way the column name is extracted from the `column` string.